Add advanced session duration filter option

+++ backend/private-graph/graph/model/models_gen.go
@@ -52,8 +52,8 @@ type ErrorTrace struct {
 }
 
 type LengthRangeInput struct {
-	Min *int `json:"min"`
-	Max *int `json:"max"`
+	Min *float64 `json:"min"`
+	Max *float64 `json:"max"`
 }
 
 type NewUsersCount struct {
